Added add_context generic functions to add to call movies and tvshow constructors. Added tests

+++ Pholos/include/database.hpp
@@ -35,7 +35,7 @@ class Database {
   void create_tvshow_table();
   void create_season_table();
 
-  int get_element_id(const std::string &name, const char flag) const;
+  unsigned long long get_element_id(const std::string &name, const char flag) const;
   void Database::add_season(const std::string &name, const std::map<int, int> &season);
 
  private:
